fix compile issues really this time
